# xWindowsOptionalFeature: Suppress useless verbose output