@mhendric 2018-11-04
Question for @mkht , and for the community at large. Can we envision any scenarios where someone would actually want to see this verbose Import-Module output? Would it ever be useful for troubleshooting (like if the Import failed for some reason)? If so, this may be better implemented as a configurable parameter to the module, not as a change that just turns it off permanently (although the new parameter could default to suppressing the output).