@mkht 2018-11-04
Thank you for reply @mhendric 

I can not imagine the situation where the output of Import-Module is useful for troubleshooting, but if we have such a scenario, we may leave a verbose message. Or how about outputting to a debug message instead of a verbose message?

In any case, I do not agree to add new parameters to determine whether to output message or not. It will make more complicated than necessary and I do not know another DSC resources that have such parameters.