@mhendric 2018-11-04
Hi @mkht , the Debug option may be a good option if it's possible. Not sure if this will work, but you could try adding -Debug:$DebugPreference to the Import-Module command, and then confirm what it outputs if you pass the -Debug switch to Start-DscConfiguration. Also confirm whether it adds all the output you're trying to suppress, so that the output is still available if someone needs it.