@mhendric 2019-01-24
I just tested, and using -Debug:$DebugPreference does not add the output lost by setting -Verbose:$false. @PlagueHO , can you weigh in on whether you think we should leave this Verbose Import-Module output or take it out? I'm fine either way.