@PlagueHO 2019-01-25
@mhendric, @mkht 

I don't recall a situation where I've found the verbose output of Import-Module useful and anytime I was trying to diagnose module load issues I was already getting an exception thrown from the module import. So the verbose logs didn't help.

That said, I generally don't apply DSC configs with Verbose enabled. I only enable it when I'm debugging something.

I'm not adverse to using `Import-Module ... -Verbose:$false` though if it adds no value and just clutters up resources. I don't recall seeing any reports of people having issues with dependent module loading (that didn't already show up as an exception in the resource).

@johlju - from memory SqlServerDsc uses Import-Module - what do you do over there?