@mkht 2019-01-26
Thanks @mhendric, @PlagueHO 

I will immediately update the document and prepare to merge it into the dev branch.

----
For your reference, 
If you want to output the output of Import-Module to Debug stream instead of Verbose, you can do with such code.

```PowerShell
    # Check that Dism PowerShell module is available
    Import-Module -Name 'Dism' -ErrorVariable 'errorsFromDismImport' -ErrorAction 'SilentlyContinue' -Force -Verbose 4>&1 | Write-Debug
```