@johlju 2019-01-26
In the SqlServerDsc module we have a helper function (`Import-SQLPSModule`) that imports the correct module (since we can either use SQLPS module or SqlServer module), but the `Import-Module` is using `-Verbose:$false` because it would otherwise spam the verbose logs with non-relevant information (in our case non-relevant at least).