@PlagueHO 2019-01-26
@mkht - maybe we just go with what @johlju does over in SQLServer? No need to move it out into a wrapper function (as @johlju mentioned they have additional logic to care for). Seems like it would be the sensible approach to replicate that.