@mhendric 2019-01-26
I think what @johlju suggested (minus the helper function) is what @mkht already did with this submission (set -Verbose:$false). If we're all in agreement that that's the right approach, I think we just need to resolve conflicts and call it good. Ultimately I could still go either way on this, but don't think it's worth spending a ton of time debating this one. I don't necessarily like removing logging, but I agree that this particular logging is probably never gonna be useful (and if there are issues, they will probably show up as a Warning or Error).