@PlagueHO 2019-01-26
So I guess we go with what @mkht said from the get-go :grin: . Yep, better fix the readme - good catch!